+++ analysis_options.yaml
@@ -24,6 +24,8 @@ linter:
   rules:
     # avoid_print: false  # Uncomment to disable the `avoid_print` rule
     # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
+    prefer_const_constructors: false
+    prefer_const_literals_to_create_immutables: false
 
 # Additional information about this file can be found at
 # https://dart.dev/guides/language/analysis-options
